refactor: remove HeroSocialLinks component and restructure files
- Removed `HeroSocialLinks` component and its associated test file.
- Updated `Hero` component to reflect the removal of `HeroSocialLinks`.
- Reorganised `HeroImage` and `HeroIntro` into dedicated subdirectories.
- Renamed `HeroImage.tsx` to `HeroImage/HeroImage.tsx`.
- Renamed `HeroIntro.tsx` to `HeroIntro/HeroIntro.tsx`.
- Updated alt text for `HeroImage` to improve clarity.
